requirements.txt
datasets==2.18.0
difPy==4.1.0
difPy==4.1.0
fastai==2.7.14
matplotlib==3.7.2
pandas==1.5.3
Pillow==10.3.0
scikit_learn==1.3.0
timm==0.9.16
torch==2.1.0.dev20230704+cu121
tqdm==4.65.0
